fix(node): validate a hosted cluster ID where a cluster is supplied, not where a behavior is built
The check sat in ClusterBehaviorType, which a peer's cluster also passes
through. The base type for a discovered cluster is built there without
forClient, so a peer whose ID resolves against a custom Matter model reached the
check and node initialization still failed — the crash this branch set out to
fix, one path over.

ClusterBehavior.for() is the entry point through which a developer supplies a
cluster to host, and every other route into ClusterBehaviorType either starts
from an already validated base or belongs to a peer. Moving the check there
covers what it is meant to cover and leaves peer behavior generation alone.

The diagnostic claimed a vendor-specific cluster may use any non-zero prefix, but
a prefix above 0xfff4 is not a legal identifier either, so an ID rejected for its
prefix was described as satisfying the rule. It now states both ranges.
